move AzureDirectorySearchService to a library
so MVC can keep older json version needed for lookups
