+++ Purchasing.Mvc/packages.config
@@ -2,7 +2,6 @@
 <packages>
   <package id="Antlr" version="3.5.0.2" targetFramework="net45" />
   <package id="AutoMapper" version="3.3.1" targetFramework="net451" />
-  <package id="AzureActiveDirectorySearcher" version="0.6.0.12" targetFramework="net452" />
   <package id="bootstrap" version="3.3.0" targetFramework="net45" />
   <package id="Castle.Core" version="3.2.0" targetFramework="net45" />
   <package id="Castle.Windsor" version="3.2.1" targetFramework="net45" />
@@ -24,11 +23,9 @@
   <package id="Microsoft.AspNet.Razor" version="3.2.2" targetFramework="net45" />
   <package id="Microsoft.AspNet.Web.Optimization" version="1.1.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.2" targetFramework="net45" />
-  <package id="Microsoft.Azure.ActiveDirectory.GraphClient" version="2.1.1" targetFramework="net452" />
   <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net452" />
   <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net452" />
   <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net452" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.13.8" targetFramework="net452" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.2" targetFramework="net45" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
   <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.1.0" targetFramework="net451" />
@@ -36,7 +33,7 @@
   <package id="Modernizr" version="2.8.3" targetFramework="net45" />
   <package id="MvcContrib.Mvc3.FluentHtml-ci" version="3.0.100.0" targetFramework="net45" />
   <package id="NEST" version="1.3.1" targetFramework="net451" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="6.0.6" targetFramework="net452" />
   <package id="NHibernate" version="3.3.1.4000" targetFramework="net45" />
   <package id="NPOI" version="2.0.6" targetFramework="net451" />
   <package id="OctoPack" version="3.0.71" targetFramework="net451" developmentDependency="true" />
