+++ Purchasing.Service/AzureDirectorySearchService.cs
@@ -6,6 +6,34 @@ using Microsoft.Azure;
 
 namespace Purchasing.Mvc.Services
 {
+    public class DirectoryUser
+    {
+        public string EmployeeId { get; set; }
+        public string LoginId { get; set; }
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public string FullName { get; set; }
+        public string EmailAddress { get; set; }
+        public string PhoneNumber { get; set; }
+    }
+
+    public interface IDirectorySearchService
+    {
+        /// <summary>
+        /// Searches for users across many different critera
+        /// </summary>
+        /// <param name="searchTerm">
+        /// Login, email or lastName
+        /// </param>
+        /// <returns></returns>
+        List<DirectoryUser> SearchUsers(string searchTerm);
+
+        /// <summary>
+        /// Returns the single user that matches the search term -- either loginID or email
+        /// </summary>
+        DirectoryUser FindUser(string searchTerm);
+    }
+
     public class AzureDirectorySearchService : IDirectorySearchService
     {
         private readonly GraphSearchClient _searcher;
