+++ Purchasing.Service/packages.config
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="AzureActiveDirectorySearcher" version="0.6.0.12" targetFramework="net452" />
+  <package id="Microsoft.Azure.ActiveDirectory.GraphClient" version="2.1.1" targetFramework="net452" />
+  <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net452" />
+  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.13.8" targetFramework="net452" />
+  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.1.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
+  <package id="System.Spatial" version="5.6.4" targetFramework="net452" />
+</packages>
