+++ Purchasing.sln
@@ -36,6 +36,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Purchasing.Jobs.DailyEmailN
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Purchasing.Jobs.NotificationsCommon", "Purchasing.Jobs.NotificationsCommon\Purchasing.Jobs.NotificationsCommon.csproj", "{029BA1FC-A03F-490D-AF92-FA69E6565ABA}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Purchasing.Service", "Purchasing.Service\Purchasing.Service.csproj", "{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -172,6 +174,16 @@ Global
 		{029BA1FC-A03F-490D-AF92-FA69E6565ABA}.ReleaseCAES|Any CPU.Build.0 = Release|Any CPU
 		{029BA1FC-A03F-490D-AF92-FA69E6565ABA}.Test|Any CPU.ActiveCfg = Release|Any CPU
 		{029BA1FC-A03F-490D-AF92-FA69E6565ABA}.Test|Any CPU.Build.0 = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.ReleaseAzure|Any CPU.ActiveCfg = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.ReleaseAzure|Any CPU.Build.0 = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.ReleaseCAES|Any CPU.ActiveCfg = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.ReleaseCAES|Any CPU.Build.0 = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Test|Any CPU.ActiveCfg = Release|Any CPU
+		{057AF0B5-BDE2-4837-9A29-8D73091DE6A6}.Test|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
